moved CapabilitiesFactory
