+++ src/main/java/util/driver/CapabilitiesFactory.java
@@ -1,130 +1,21 @@
 package util.driver;
 
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.phantomjs.PhantomJSDriverService;
 import org.openqa.selenium.remote.DesiredCapabilities;
 
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
 
-import static net.itarray.automotion.tools.environment.EnvironmentFactory.*;
-
+/**
+ * @deprecated As of release 2.0, replaced by{@link net.itarray.automotion.tools.util.driver.CapabilitiesFactory}
+ */
 public class CapabilitiesFactory {
 
     private static DesiredCapabilities capabilities = new DesiredCapabilities();
 
     public static DesiredCapabilities getCapabilities() {
-        if (isAndroid()) {
-            capabilities = getAndroidCapabilities();
-        } else if (isIOS()) {
-            capabilities = getIOSCapabilities();
-        } else if (isWindows()) {
-            capabilities = getWindowsCapabilities();
-        } else if (isRemote()) {
-            capabilities = getRemoteDriverCapabilities();
-        } else if (isHeadless()) {
-            capabilities = getPhantomJSCapabilities();
-        }
-
-        return capabilities;
+        return net.itarray.automotion.tools.util.driver.CapabilitiesFactory.getCapabilities();
     }
 
     public static DesiredCapabilities updateCapabilities(DesiredCapabilities desiredCapabilities, Map<String, Object> mapCapabilities) {
-        if (mapCapabilities.size() > 0) {
-            for (Map.Entry<String, Object> capability : mapCapabilities.entrySet()) {
-                desiredCapabilities.setCapability(capability.getKey(), capability.getValue());
-            }
-        }
-
-        return desiredCapabilities;
-    }
-
-    private static DesiredCapabilities getCommonMobileCapabilities() {
-        capabilities.setCapability("automationName", getAutomationName());
-        capabilities.setCapability("platformVersion", getPlatformVersion());
-        capabilities.setCapability("deviceName", getDevice());
-        capabilities.setCapability("deviceOrientation", "portrait");
-        capabilities.setCapability("app", getApp());
-        capabilities.setCapability("browserName", getMobileBrowser());
-        capabilities.setCapability("appiumVersion", getAppiumVersion());
-        capabilities.setCapability("name", getName());
-        capabilities.setCapability("newCommandTimeout", getNewCommandTimeout());
-
-        return capabilities;
-    }
-
-    private static DesiredCapabilities getAndroidCapabilities() {
-        capabilities = getCommonMobileCapabilities();
-        capabilities.setCapability("platformName", "Android");
-        capabilities.setCapability("appActivity", getAppActivity());
-        capabilities.setCapability("appPackage", getAppPackage());
-
-        return capabilities;
-    }
-
-    private static DesiredCapabilities getIOSCapabilities() {
-        capabilities = getCommonMobileCapabilities();
-        capabilities.setCapability("platformName", "iOS");
-        capabilities.setCapability("udid", getUDIDDevice());
-        capabilities.setCapability("waitForAppScript", "true");
-
-        return capabilities;
-    }
-
-    private static DesiredCapabilities getWindowsCapabilities() {
-        capabilities = getCommonMobileCapabilities();
-        capabilities.setCapability("platformName", "Windows");
-
-        return capabilities;
-    }
-
-    private static DesiredCapabilities getPhantomJSCapabilities() {
-        capabilities.setJavascriptEnabled(true);
-        capabilities.setCapability("takesScreenshot", true);
-        capabilities.setCapability("browserName", "PhantomJS");
-        capabilities.setCapability("browser", "phantomjs");
-        capabilities.setCapability("phantomjs.page.settings.userAgent", getUserAgent());
-        capabilities.setCapability("acceptSslCerts", true);
-        capabilities.setCapability(PhantomJSDriverService.PHANTOMJS_CLI_ARGS, new String[] {"--web-security=no", "--ignore-ssl-errors=yes","--ignore-ssl-errors=true","--ssl-protocol=tlsv1"});
-        capabilities.setCapability(PhantomJSDriverService.PHANTOMJS_EXECUTABLE_PATH_PROPERTY,
-                getPhantomJsPath()
-        );
-
-        return capabilities;
-    }
-
-    private static DesiredCapabilities getRemoteDriverCapabilities() {
-        if (isFirefox()) {
-            capabilities = DesiredCapabilities.firefox();
-        } else if (isChrome()) {
-            capabilities = DesiredCapabilities.chrome();
-        } else if (isSafari()) {
-            capabilities = DesiredCapabilities.safari();
-        } else if (isInternetExplorer()) {
-            capabilities = DesiredCapabilities.internetExplorer();
-        } else if (isEDGE()) {
-            capabilities = DesiredCapabilities.edge();
-        }
-
-        if (getMobileDeviveEmulation() != null) {
-            Map<String, String> mobileEmulation = new HashMap<>();
-            mobileEmulation.put("deviceName", getMobileDeviveEmulation());
-
-            Map<String, Object> chromeOptions = new HashMap<>();
-            chromeOptions.put("mobileEmulation", mobileEmulation);
-            chromeOptions.put("args", Arrays.asList("enable-extensions",
-                    "test-type", "no-default-browser-check", "ignore-certificate-errors"));
-
-            capabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
-        }
-
-        if (getSlDesktopPlatform() != null) {
-            capabilities.setCapability("platform", getSlDesktopPlatform());
-            capabilities.setCapability("version", getSlBrowserVersion());
-            capabilities.setCapability("screenResolution", getSlDesktopResolution());
-        }
-
-        return capabilities;
+        return net.itarray.automotion.tools.util.driver.CapabilitiesFactory.updateCapabilities(desiredCapabilities, mapCapabilities);
     }
 }
