+++ src/main/java/util/driver/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * @deprecated As of release 2.0, replaced by{@link net.itarray.automotion.tools.util.driver}
+ */
+package util.driver;
